requirements.txt
streamlit==0.82.0
ToushinReader==0.2.3
altair==4.1.0
python-dateutil==2.8.2
pandas-datareader==0.10.0
requests-oauthlib==1.3.0
widgetsnbextension==3.5.1
